Properly handle invalid viewports

+++ src/gpu_3d.h
@@ -188,8 +188,8 @@ class Gpu3D
         uint32_t lightColor[4] = {};
         uint8_t shininess[128] = {};
 
-        int viewportX = 0, viewportY = 0;
-        int viewportWidth = 256, viewportHeight = 192;
+        uint16_t viewportX = 0, viewportY = 0;
+        uint16_t viewportWidth = 256, viewportHeight = 192;
 
         int16_t boxTestCoords[6] = {};
